swedish_multimodal/requirements.txt
openai
PyMuPDF
python-dotenv
datasets